Refactor cart functionality; remove unused checkout and product listing HTML files, implement ShoppingCart class for better cart management
