--- src/cart/checkout.html
@@ -1,23 +1,0 @@
-<!doctype html>
-
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-    <title>Sleep Outside | Cart</title>
-
-    <link rel="stylesheet" href="/css/style.css" />
-
-    <script src="../js/checkout.js" type="module"></script>
-  </head>
-
-  <body>
-    <header class="divider" id="main-header"></header>
-
-    <main class="divider"></main>
-
-    <footer id="main-footer"></footer>
-  </body>
-</html>
